fix the CI issue
