+++ cypress/tsconfig.json
@@ -3,6 +3,7 @@
   "exclude": [],
   "compilerOptions": {
     "types": ["cypress", "@testing-library/cypress"],
+    "target": "ES6",
     "isolatedModules": false,
     "noEmit": true
   }
