+++ src/Routes.tsx
@@ -21,17 +21,6 @@ import {
   Projects,
 } from './pages';
 
-export const AppRoute = {
-  index: '/',
-  auth: '/auth',
-  login: '/auth/login',
-  dashboard: '/dashboard',
-  families: '/dashboard/families',
-  familyDetail: '/dashboard/families/:id',
-  projects: '/dashboard/projects',
-} as const;
-export type AppRoute = typeof AppRoute[keyof typeof AppRoute];
-
 export const location = new ReactLocation();
 
 export type LocationGenerics = MakeGenerics<{
