+++ src/components/CreateFamily/FamilyForm/CreateFamilyForm.ids.ts
@@ -1,0 +1,9 @@
+export const createFamilyFormIds = {
+  form: 'create-family-form',
+  nameInput: 'family-name',
+  submitBtn: 'submit-button',
+  notification: {
+    success: 'create-family-success-notification',
+    failure: 'create-family-failure-notification',
+  },
+} as const;
