+++ src/components/CreateProject/ProjectForm/CreateProjectForm.ids.ts
@@ -1,0 +1,10 @@
+export const createProjectFormIds = {
+  form: 'create-project-form',
+  nameInput: 'project-name',
+  descriptionInput: 'project-description',
+  submitBtn: 'submit-button',
+  notification: {
+    success: 'create-project-success-notification',
+    failure: 'create-project-failure-notification',
+  },
+} as const;
