+++ src/components/LogoutModal/LogoutModal.ids.ts
@@ -1,0 +1,4 @@
+export const logoutModalIds = {
+  modal: 'logout-modal',
+  acceptBtn: 'logout-modal-accept-button',
+};
