+++ src/components/LogoutModal/LogoutModal.tsx
@@ -4,12 +4,9 @@ import type { ModalProps } from '@mantine/core';
 import { Button, Group, Modal, Stack, Text } from '@mantine/core';
 
 import { createTestAttr } from '../../utils/createTestAttr';
+import { logoutModalIds as ids } from './LogoutModal.ids';
 
 type Props = Pick<ModalProps, 'onClose' | 'opened'>;
-export const logoutModalIds = {
-  modal: 'logout-modal',
-  acceptBtn: 'logout-modal-accept-button',
-};
 const texts = messages.logout.modal;
 
 export const LogoutModal = ({ opened, onClose }: Props) => {
@@ -20,7 +17,7 @@ export const LogoutModal = ({ opened, onClose }: Props) => {
       centered
       opened={opened}
       onClose={onClose}
-      {...createTestAttr(logoutModalIds.modal)}
+      {...createTestAttr(ids.modal)}
     >
       <Stack spacing={40}>
         <Text color="fgDefault">{texts.text}</Text>
@@ -28,7 +25,7 @@ export const LogoutModal = ({ opened, onClose }: Props) => {
           <Button
             variant="filled"
             color="red"
-            {...createTestAttr(logoutModalIds.acceptBtn)}
+            {...createTestAttr(ids.acceptBtn)}
             onClick={() => {
               logoutLocally();
             }}
