+++ src/components/atoms/NavLink/ExitNavLink.ids.ts
@@ -1,0 +1,1 @@
+export const exitNavLinkId = 'nav-link-exit';
