+++ src/components/organisms/Header/DashboardHeader.ids.ts
@@ -1,0 +1,1 @@
+export const dashboardHeaderId = 'dashboard-header';
