+++ src/components/organisms/Sidebar/Sidebar.ids.ts
@@ -1,0 +1,4 @@
+export const navLinkIds = {
+  families: 'nav-link-families',
+  projects: 'nav-link-projects',
+};
