+++ src/components/organisms/Sidebar/Sidebar.stories.tsx
@@ -1,6 +1,6 @@
 import type { Meta, Story } from '@storybook/react';
 
-import { SideBar } from './SideBar';
+import { SideBar } from './Sidebar';
 
 export default {
   component: SideBar,
