+++ src/components/organisms/Sidebar/index.ts
@@ -1,1 +1,1 @@
-export * from './SideBar';
+export * from './Sidebar';
