+++ src/utils/createTestAttr.ts
@@ -1,6 +1,3 @@
-const dataAttrName = 'testid';
-
-export function createTestAttr(id: string, type: 'data-' | 'id' = 'data-') {
-  if (type === 'id') return { id };
-  return { [`data-${dataAttrName}`]: id };
+export function createTestAttr(id: string) {
+  return { [`data-testid`]: id };
 }
